WIP: refactor: Introduce "cache-control" plugin and switch req. pipeline to it.
> This commit message is WIP.  The commit is "done".

Simialr to f3efe04b5c3, which migrated the tracing package to the plugin
API, this commit introduces a `plugin` (named) export from the
`apollo-cache-control` module.  It similarly accomplishes that by
duplicating the behavior of the `CacheControlExtension` which leveraged the
soon-to-be-deprecated `graphql-extensions`.  The functionality, again, is
intended to be identical in spirit.

Since the delta of the commits was otherwise even _more_ confusing, I've
left the `CacheControlExtension` present and exported and will remove it in a
subsequent commit.

Briefly summarizing what the necessary changes were:

1. We no longer use a class instance to house the extension, which was
   necessitated by the `graphql-extensions` API.  This means that uses of
   `this` have been replaced with function scoped variables by the same name.
2. The logic which actually does the formatting (previously handled by the
   `format` method in `graphql-extension`, now takes place within the plugin
   API's `willSendResponse` method.
3. *overallCachePolicy*
4. anything else?